Preferences: Move ColorSchemeConfigPage to its own module and rename it AppearanceConfigPage
